% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMO.R
\name{GetMO}
\alias{GetMO}
\title{GetMO}
\usage{
GetMO(
  INPUT,
  CELL_ID_COLUMN,
  X_POSITION,
  Y_POSITION,
  ANNO_COLUMN,
  RADIUS = "auto",
  MEDIAN_NB = 15,
  ORIENTEDNESS = T,
  VERBOSE = 1
)
}
\arguments{
\item{INPUT}{A data frame that must contain the image ID, cell ID, the coordinates of the cell in the image, and the cell type (tumor or non-tumor).}

\item{CELL_ID_COLUMN}{The cell ID.}

\item{X_POSITION}{The x-axis coordinate of cell in image.}

\item{Y_POSITION}{The y-axis coordinate of cell in image.}

\item{ANNO_COLUMN}{The cell type, binary classification (tumor -> 1 or non-tumor -> 0).}

\item{RADIUS}{Default 'auto'.}

\item{MEDIAN_NB}{Default 15. Recommended 15 to 30.}

\item{ORIENTEDNESS}{TRUE or FALSE. Default T.}

\item{VERBOSE}{Default 1.}
}
\value{
A dataframe that contain tumor boundary and the microenvironment inside and outside the tumor divided by the boundary.
}
\description{
This function takes a name and returns a greeting message.
}
\examples{
GetMO(CELL_ID_COLUMN = 'Cell_ID', X_POSITION = 'X_position', Y_POSITION = 'Y_position', ANNO_COLUMN = 'CT0', RADIUS = 'auto', MEDIAN_NB = 15, ORIENTEDNESS = F, VERBOSE = 1)
}
