% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDist2Boundary.R
\name{GetDist2Boundary}
\alias{GetDist2Boundary}
\title{GetDist2Boundary}
\usage{
GetDist2Boundary(
  INPUT,
  CELL_ID_COLUMN,
  X_POSITION,
  Y_POSITION,
  ANNO_COLUMN,
  ANNO_OF_BOUNDARY,
  K = 5
)
}
\arguments{
\item{INPUT}{A data frame that must contain the image ID, cell ID, the coordinates of the cell in the image, and the cell type (tumor or non-tumor).}

\item{CELL_ID_COLUMN}{The cell ID.}

\item{X_POSITION}{The x-axis coordinate of cell in image.}

\item{Y_POSITION}{The y-axis coordinate of cell in image.}

\item{ANNO_COLUMN}{The cell type, binary classification (tumor -> 1 or non-tumor -> 0).}

\item{ANNO_OF_BOUNDARY}{The tumor boundary annotation derived from GetBoundary.}

\item{K}{Default 5.}
}
\value{
A dataframe that contain the microenvironment inside and outside the tumor divided by the boundary.
}
\description{
This function takes a name and returns a greeting message.
}
\examples{
GetDist2Boundary(CELL_ID_COLUMN = 'Cell_ID', X_POSITION = 'X_position', Y_POSITION = 'Y_position', ANNO_COLUMN = 'CT0_2', ANNO_OF_BOUNDARY = 'Boundary')
}
